Projeto Novo/requirements.txt
Flask==3.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.0
